Changed to utilize API to Upload straight from conversion.
Strava still has a bug in the API where the Name and Activity type aren't recognized in the API, even though this app sets them right according to the API Spec. So you have to rename your activity if you don't like hte date as the name, and change the type if it's not a ride.