Tons of changes. New JAR File.
Added the ability to upload to RideWithGPS as well.
Moved code around to modularize site publishing. Converter now just sucks in the file and builds an internal trackpoint list that's then published to objects that handle conversion and upload.
Fixed it to Mio support is transparent, no need to specify Device. Still "forges" it as a Garmin 800 though to force elevation data to be used correctly. This method should make it easier to add new names for the same data down the road.
App now uses the Registry to remember settings, so you don't have to hit every button every time you use it. Also remembers the last file location when you browse for the file to upload.
Only one credential box, so if you upload to both RWGPS and Strava, the logins need to match. Otherwise, edit the values, check one, do the upload, then change the values, check the other, and upload again.
So many changes, I will not be surprised to find bugs.